refactor(graduate): 공지사항 관련 API 및 훅 사용 도메인으로 위치 변경
