+++ apps/graduate/src/admin/pages/notice/api/index.ts
@@ -1,0 +1,4 @@
+export * from './removeNotice';
+export * from './submitNotice';
+export * from './submitNoticeFile';
+export * from './updateNotice';
