--- apps/graduate/src/admin/pages/notice/hooks/index.ts
@@ -1,3 +1,0 @@
-export * from './useCreateNotice';
-export * from './useUpdateNotice';
-export * from './useDeleteNotice';
