+++ apps/graduate/src/admin/pages/notice/model/index.ts
@@ -1,1 +1,4 @@
 export type * from './notices';
+export * from './useCreateNotice';
+export * from './useUpdateNotice';
+export * from './useRemoveNotice';
