+++ apps/graduate/src/admin/shared/api/index.ts
@@ -1,2 +1,1 @@
 export * from './schedule';
-export * from './notice';
